requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
httpcore==0.17.3
